Ensure enums are always exported
This lets us identify type-only imports so Jest doesn’t get confused by the built Babel output
