+++ model/src/components/enums.ts
@@ -1,0 +1,26 @@
+export enum ComponentTypeEnum {
+  TextField = 'TextField',
+  MultilineTextField = 'MultilineTextField',
+  YesNoField = 'YesNoField',
+  DateField = 'DateField',
+  TimeField = 'TimeField',
+  DateTimeField = 'DateTimeField',
+  DatePartsField = 'DatePartsField',
+  MonthYearField = 'MonthYearField',
+  DateTimePartsField = 'DateTimePartsField',
+  SelectField = 'SelectField',
+  AutocompleteField = 'AutocompleteField',
+  RadiosField = 'RadiosField',
+  CheckboxesField = 'CheckboxesField',
+  NumberField = 'NumberField',
+  UkAddressField = 'UkAddressField',
+  TelephoneNumberField = 'TelephoneNumberField',
+  EmailAddressField = 'EmailAddressField',
+  FileUploadField = 'FileUploadField',
+  Para = 'Para',
+  Html = 'Html',
+  InsetText = 'InsetText',
+  Details = 'Details',
+  FlashCard = 'FlashCard',
+  List = 'List'
+}
