+++ model/src/components/index.ts
@@ -1,3 +1,4 @@
 export { ComponentTypes } from './component-types'
 export { ConditionalComponentTypes } from './conditional-component-types'
-export * from './types'
+export { ComponentTypeEnum } from './enums'
+export type * from './types'
