+++ model/src/conditions/enums.ts
@@ -1,0 +1,4 @@
+export enum Coordinator {
+  AND = 'and',
+  OR = 'or'
+}
