+++ model/src/conditions/types.ts
@@ -4,11 +4,6 @@ import { ConditionGroup } from './condition-group'
 
 export type ConditionsArray = (Condition | ConditionGroup | ConditionRef)[]
 
-export enum Coordinator {
-  AND = 'and',
-  OR = 'or'
-}
-
 export type DateTimeUnitValues =
   | 'years'
   | 'months'
