+++ model/src/data-model/enums.ts
@@ -1,0 +1,5 @@
+export enum OutputType {
+  Email = 'email',
+  Notify = 'notify',
+  Webhook = 'webhook'
+}
