+++ model/src/data-model/index.ts
@@ -1,3 +1,4 @@
 export { InputWrapper } from './input-wrapper'
 export { ConditionsWrapper, ConditionRawData } from './conditions-wrapper'
 export { Page, Item, Section, List, ConfirmationPage } from './types'
+export { OutputType } from './enums'
